Fixed cyclical dependency
