+++ generator/typescript/config.go
@@ -75,5 +75,3 @@ func (c *Config) AddCustomType(name, value string) {
 func (c *Config) Generator() *Generator {
 	return NewGenerator(*c)
 }
-
-var _ config.TargetInterface = &Config{}
